updatsed RiemannianDistance
